requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
Pillow>=10.0.0
scipy>=1.11.0
